+++ shaders/grass_f.glsl
@@ -18,7 +18,7 @@ uniform vec4 fog;
 
 varying vec4 pointLight [10];
 uniform vec4 light_color[10];
-uniform int num_lights;
+uniform float num_lights;
 varying vec4 vpos;
 void main()
     {    
@@ -69,7 +69,7 @@ void main()
         //point lights                 
         //float att;
         //float dist;
-        //for (int i=0; i<num_lights; ++i)
+        //for (int i=0; i<int(num_lights); ++i)
         //    {  
         //    dist=dist=distance(vpos.xyz, pointLight[i].xyz);
         //    dist*=dist;
